requirements.txt
streamlit
openai==0.28.0
python-dotenv
pandas
globus_sdk 
firebase-admin
python-docx==0.8.11
pdf2image
Pillow
PyPDF2
pdfrw
